Move vis_type_xy into NP
