--- src/legacy/core_plugins/vis_type_xy/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "visTypeXy",
-  "version": "kibana"
-}
